+++ app/controllers/application_controller.rb
@@ -1,7 +1,20 @@
 class ApplicationController < ActionController::Base
+  before_action :configure_permitted_parameters, if: :devise_controller?
+
   # Prevent CSRF attacks by raising an exception.
   # For APIs, you may want to use :null_session instead.
   protect_from_forgery with: :exception
 
   layout 'conductor-application-v2'
+
+  protected
+
+  def configure_permitted_parameters
+    devise_parameter_sanitizer.for(:sign_up) { |u| u.permit(:first_name, :last_name, :email, :password, :password_confirmation) }
+  end
+
+  def check_system_admin
+    redirect_to root_path, alert: "You do not have sufficient privileges to access that page." unless current_user.system_admin?
+  end
+
 end