+++ db/migrate/20150614180423_create_projects.rb
@@ -1,0 +1,13 @@
+class CreateProjects < ActiveRecord::Migration
+  def change
+    create_table :projects do |t|
+      t.string :name
+      t.string :slug
+      t.string :logo
+      t.integer :user_id
+      t.boolean :deleted, null: false, default: false
+
+      t.timestamps null: false
+    end
+  end
+end