+++ test/test_helper.rb
@@ -11,3 +11,34 @@ class ActiveSupport::TestCase
 
   # Add more helper methods to be used by all tests here...
 end
+
+class ActionController::TestCase
+  include Devise::TestHelpers
+
+  def login(resource)
+    @request.env["devise.mapping"] = Devise.mappings[resource]
+    sign_in(resource.class.name.downcase.to_sym, resource)
+  end
+end
+
+class ActionDispatch::IntegrationTest
+  def sign_in_as(user_template, password, email)
+    user = User.create(password: password, password_confirmation: password, email: email,
+                       first_name: user_template.first_name, last_name: user_template.last_name)
+    user.save!
+    user.update_column :deleted, user_template.deleted?
+    user.update_column :system_admin, user_template.system_admin?
+    post_via_redirect '/login', user: { email: email, password: password }
+    user
+  end
+end
+
+module Rack
+  module Test
+    class UploadedFile
+      def tempfile
+        @tempfile
+      end
+    end
+  end
+end